bugfix: 도감 조회 시 타입 불일치 문제 해결

See merge request s13-webmobile1-sub1/S13P11C106!102

+++ BE/StellarVision/src/main/java/com/susang/stellarVision/application/collection/repository/MemberCollectionRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MemberCollectionRepository extends JpaRepository<MemberCollection, Long> {
     @Query("SELECT mc.collection.id FROM MemberCollection mc WHERE mc.member.id = :memberId")
-    Set<Long> findCollectionIdsByMemberId(@Param("memberId") Long memberId);
+    Set<Integer> findCollectionIdsByMemberId(@Param("memberId") Long memberId);
 
     List<MemberCollection> findByMemberId(Long memberId);
